// src/notification/types/notification-type.enum.ts
export enum NotificationType {
    GLOBAL = "GLOBAL",
    PERSONAL = "PERSONAL",
    PAYMENT = "PAYMENT",
    VERIFICATION = "VERIFICATION",
    VERIFICATION_REFUSED = "VERIFICATION_REFUSED",
    ARTISAN_CARD = "ARTISAN_CARD",
    ARTISAN_CARD_REFUSED = "ARTISAN_CARD_REFUSED",
    HIRING = "HIRING",
    WARNING = "WARNING"
}